+++ include/hashset.h
@@ -10,7 +10,7 @@ typedef struct {
     u32 capacity;
     u32 size;
 
-    u32 elm_size;
+    u16 elm_size;
 
     custom_hash_fn hash_fn;
     delete_fn elm_del_fn;
@@ -18,7 +18,7 @@ typedef struct {
 } __attribute__((aligned(64))) hashset;
 
 
-hashset* hashset_create(u32 elm_size, custom_hash_fn hash_fn, 
+hashset* hashset_create(u16 elm_size, custom_hash_fn hash_fn, 
                         delete_fn elm_del, compare_fn cmp_fn);
 
 void hashset_destroy(hashset* set);
